+++ src/main/java/Kon/models/consoleSales/client/ConsoleSalesRequest.java
@@ -5,7 +5,7 @@ public class ConsoleSalesRequest {
     private String consoleName;
     private String manufacturer;
     private int releaseYear;
-    private float sales;
+    private double sales;
 
 
     public String getConsoleId() {
@@ -40,11 +40,11 @@ public class ConsoleSalesRequest {
         this.releaseYear = releaseYear;
     }
 
-    public float getSales() {
+    public double getSales() {
         return sales;
     }
 
-    public void setSales(float sales) {
+    public void setSales(double sales) {
         this.sales = sales;
     }
 }
